Move update button to controlbar and improve the update screen
