+++ lyrebird/src/main/java/moe/lyrebird/view/components/controlbar/ControlBar.fxml
@@ -53,6 +53,7 @@
                   <Button fx:id="mentionsViewButton" maxWidth="1.7976931348623157E308" mnemonicParsing="false" text="Mentions" />
                   <Button fx:id="directMessagesViewButton" maxWidth="1.7976931348623157E308" mnemonicParsing="false" text="DM" />
                   <Button fx:id="creditsButton" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" mnemonicParsing="false" text="About" />
+                  <Button fx:id="updateButton" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" mnemonicParsing="false" text="Update" managed="false" visible="false" />
                </children>
             </VBox>
          </center>
